requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
common==0.1.2
marshmallow==3.13.0
flask_marshmallow==0.14.0
get_project_root==0.2
great-expectations==0.13.35
pyspark==3.1.2
pyarrow==3.0.0
ruamel.yaml==0.17.16
pandas==1.3.3
google-api-python-client==2.24.0
google-cloud-bigquery==2.28.0
google-cloud-storage==1.42.3
sqlalchemy==1.4.25
pybigquery==0.10.2
google-cloud-bigquery-storage==2.9.0
get-project-root==0.2